+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.DailyPet"
         tools:targetApi="31">
-
+        <activity
+            android:name=".TestActivity"
+            android:exported="false" />
         <activity
             android:name=".LoginActivity"
             android:exported="false" />
