+++ app/src/main/res/layout/activity_login.xml
@@ -23,10 +23,10 @@
 
     <ImageButton
         android:id="@+id/btn_naverLogin"
-        android:layout_width="224dp"
-        android:layout_height="63dp"
+        android:layout_width="219dp"
+        android:layout_height="56dp"
         android:background="@android:color/transparent"
-        android:scaleType="fitCenter"
+        android:scaleType="fitXY"
         android:src="@drawable/naver_login_large_green"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
@@ -41,10 +41,24 @@
         android:id="@+id/img_loginLogo"
         android:layout_width="351dp"
         android:layout_height="342dp"
-        android:layout_marginTop="88dp"
+        android:layout_marginTop="120dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:srcCompat="@drawable/app_logo1" />
+
+    <TextView
+        android:id="@+id/text_loginSlogan"
+        android:layout_width="293dp"
+        android:layout_height="44dp"
+        android:layout_marginTop="76dp"
+        android:text="슬로건 문구"
+        android:textAlignment="center"
+        android:textColor="@color/teal_700"
+        android:textSize="20sp"
+        android:typeface="normal"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
