+++ app/src/main/res/values/colors.xml
@@ -8,6 +8,6 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="logo_white">#F2F2F2</color>
-    <color name="main_blue">#F2778D</color>
-    <color name="main_pink">#73D9D9</color>
+    <color name="main_pink">#F2778D</color>
+    <color name="main_blue">#73D9D9</color>
 </resources>
