+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-        <string name="app_name">반려하루</string>
+    <string name="app_name">반려하루</string>
 </resources>
